Conflicts:
	.gitignore
	bin/start.ps1
	jeeves/src/main/java/jeeves/server/context/BasicContext.java
	web-client/pom.xml
	web-client/src/main/resources/apps/css/gndefault.css
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/Catalogue.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/form/GeometryMapField.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/form/OpenSearchSuggestionTextField.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/lang/en.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/lang/fr.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/widgets/LoginForm.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/widgets/MetadataMenu.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/widgets/MetadataResultsToolbar.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/widgets/MetadataResultsView.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/widgets/Templates.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/widgets/editor/ConceptSelectionPanel.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/widgets/editor/EditorPanel.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/widgets/editor/EditorTools.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/widgets/editor/LinkResourcesWindow.js
	web-client/src/main/resources/apps/js/GeoNetwork/lib/GeoNetwork/widgets/editor/LinkedMetadataPanel.js
	web-client/src/main/resources/apps/js/ext-ux/DateTime/DateTime.js
	web-client/src/main/resources/apps/tabsearch/index_debug.html
	web/src/main/java/org/fao/geonet/kernel/DataManager.java
	web/src/main/java/org/fao/geonet/kernel/SchemaManager.java
	web/src/main/java/org/fao/geonet/kernel/XmlSerializer.java
	web/src/main/java/org/fao/geonet/kernel/csw/services/getrecords/SearchController.java
	web/src/main/java/org/fao/geonet/kernel/search/LuceneConfig.java
	web/src/main/java/org/fao/geonet/kernel/search/LuceneSearcher.java
	web/src/main/java/org/fao/geonet/kernel/security/ldap/GeoNetworkLDAPUserDetailsContextMapper.java
	web/src/main/java/org/fao/geonet/services/main/SearchSuggestion.java
	web/src/main/java/org/fao/geonet/services/metadata/AjaxEditUtils.java
	web/src/main/java/org/fao/geonet/services/metadata/EditUtils.java
	web/src/main/java/org/fao/geonet/services/metadata/Show.java
	web/src/main/java/org/fao/geonet/services/metadata/UpdateChildren.java
	web/src/main/java/org/fao/geonet/services/metadata/XslProcessing.java
	web/src/main/java/org/fao/geonet/util/XslUtil.java
	web/src/main/webResources/WEB-INF/web.xml
	web/src/main/webapp/WEB-INF/classes/setup/sql/migrate/v2100/migrate-default.sql
	web/src/main/webapp/WEB-INF/config-security-mapping.xml
	web/src/main/webapp/WEB-INF/config-security.xml
	web/src/main/webapp/WEB-INF/data/config/schema_plugins/iso19139/loc/cat/labels.xml
	web/src/main/webapp/WEB-INF/data/config/schema_plugins/iso19139/loc/eng/labels.xml
	web/src/main/webapp/WEB-INF/data/config/schema_plugins/iso19139/loc/ger/labels.xml
	web/src/main/webapp/WEB-INF/data/config/schema_plugins/iso19139/loc/nor/labels.xml
	web/src/main/webapp/WEB-INF/data/config/schema_plugins/iso19139/loc/rus/labels.xml
	web/src/main/webapp/WEB-INF/data/config/schema_plugins/iso19139/present/metadata-edit.xsl
	web/src/main/webapp/WEB-INF/data/config/schema_plugins/iso19139/present/metadata-view.xsl
	web/src/main/webapp/WEB-INF/data/config/schema_plugins/iso19139/process/onlinesrc-add.xsl
	web/src/main/webapp/WEB-INF/data/config/schema_plugins/iso19139/process/update-srv-attachDataset.xsl
	web/src/main/webapp/WEB-INF/data/config/schema_plugins/iso19139/update-fixed-info.xsl
	web/src/main/webapp/WEB-INF/log4j.cfg
	web/src/main/webapp/loc/eng/xml/strings.xml
	web/src/main/webapp/loc/fre/xml/strings.xml
	web/src/main/webapp/loc/spa/xml/strings.xml
	web/src/main/webapp/loc/tur/xml/config.xml
	web/src/main/webapp/loc/tur/xml/strings.xml
	web/src/main/webapp/scripts/geonetwork.js
	web/src/main/webapp/xsl/metadata/layout.xsl
	web/src/main/webapp/xsl/metadata/relation.xsl
	web/src/main/webapp/xsl/search-results-xhtml.xsl
	web/src/main/webapp/xsl/text-utilities.xsl
	web/src/test/java/org/fao/geonet/kernel/XmlSerializerTest.java

+++ web/src/main/java/org/fao/geonet/services/thumbnail/Set.java
@@ -23,20 +23,21 @@
 
 package org.fao.geonet.services.thumbnail;
 
-import jeeves.interfaces.Service;
 import jeeves.resources.dbms.Dbms;
 import jeeves.server.ServiceConfig;
 import jeeves.server.context.ServiceContext;
+import jeeves.utils.IO;
 import jeeves.utils.Util;
 import lizard.tiff.Tiff;
-
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.IOUtils;
 import org.fao.geonet.GeonetContext;
 import org.fao.geonet.constants.Geonet;
 import org.fao.geonet.constants.Params;
 import org.fao.geonet.exceptions.ConcurrentUpdateEx;
 import org.fao.geonet.kernel.DataManager;
 import org.fao.geonet.lib.Lib;
+import org.fao.geonet.services.NotInReadOnlyModeService;
 import org.jdom.Element;
 
 import javax.imageio.ImageIO;
@@ -46,10 +47,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 
-//=============================================================================
-
-public class Set implements Service
-{
+public class Set extends NotInReadOnlyModeService {
 	//--------------------------------------------------------------------------
 	//---
 	//--- Init
@@ -64,7 +62,7 @@ public class Set implements Service
 	//---
 	//--------------------------------------------------------------------------
 
-	public Element exec(Element params, ServiceContext context) throws Exception
+	public Element serviceSpecificExec(Element params, ServiceContext context) throws Exception
 	{
 		String  id            = Util.getParam     (params, Params.ID);
 		String  type          = Util.getParam     (params, Params.TYPE);
@@ -99,7 +97,7 @@ public class Set implements Service
 
 		String dataDir = Lib.resource.getDir(context, Params.Access.PUBLIC, id);
 
-		new File(dataDir).mkdirs();
+		IO.mkdirs(new File(dataDir), "Metadata data directory");
 
 		//-----------------------------------------------------------------------
 		//--- create the small thumbnail, removing the old one
@@ -110,15 +108,15 @@ public class Set implements Service
 			String inFile    = context.getUploadDir() + file;
 			String outFile   = dataDir + smallFile;
 
-			removeOldThumbnail(context, dbms, id, "small");
+			removeOldThumbnail(context, dbms, id, "small", false);
 			createThumbnail(inFile, outFile, smallScalingFactor, smallScalingDir);
-			dataMan.setThumbnail(context, dbms, id, true, smallFile);
+			dataMan.setThumbnail(context, dbms, id, true, smallFile, false);
 		}
 
 		//-----------------------------------------------------------------------
 		//--- create the requested thumbnail, removing the old one
 
-		removeOldThumbnail(context, dbms, id, type);
+		removeOldThumbnail(context, dbms, id, type, false);
 
 		if (scaling)
 		{
@@ -131,7 +129,7 @@ public class Set implements Service
 			if (!new File(inFile).delete())
 				context.error("Error while deleting thumbnail : "+inFile);
 
-			dataMan.setThumbnail(context, dbms, id, type.equals("small"), newFile);
+			dataMan.setThumbnail(context, dbms, id, type.equals("small"), newFile, false);
 		}
 		else
 		{
@@ -140,17 +138,21 @@ public class Set implements Service
 			File inFile  = new File(context.getUploadDir(), file);
 			File outFile = new File(dataDir,                file);
 
+			if(outFile.exists() && !outFile.delete()) {
+				throw new Exception("Unable to overwrite existing file: "+outFile);
+			}
 			try {
 				FileUtils.moveFile(inFile, outFile);
 			} catch (Exception e) {
-				inFile.delete();
+				IO.delete(inFile, false, context);
 				throw new Exception(
 						"Unable to move uploaded thumbnail to destination: " + outFile + ". Error: " + e.getMessage());
 			}
 
-			dataMan.setThumbnail(context, dbms, id, type.equals("small"), file);
+			dataMan.setThumbnail(context, dbms, id, type.equals("small"), file, false);
 		}
 
+		dataMan.indexInThreadPool(context, id, dbms);
 		//-----------------------------------------------------------------------
 
 		Element response = new Element("a");
@@ -160,79 +162,58 @@ public class Set implements Service
 		return response;
 	}
 
+    /**
+     * TODO javadoc.
+     *
+     * @param id
+     * @param context
+     */
+    private String createDataDir(String id, ServiceContext context) {
+        String dataDir = Lib.resource.getDir(context, Params.Access.PUBLIC, id);
+        if (!new File(dataDir).mkdirs()) {
+            context.error("Failed to make dir: " + dataDir);
+        }
+        return dataDir;
+    }
+
 	// FIXME : not elegant
 	public Element execOnHarvest(
 							Element params, 
 							ServiceContext context, 
 							Dbms dbms, 
-							DataManager dataMan) throws Exception
-	{
+							DataManager dataMan) throws Exception {
+
 		String  id            = Util.getParam     (params, Params.ID);
-		String  type          = Util.getParam     (params, Params.TYPE);
-		String  version       = Util.getParam     (params, Params.VERSION);
-		String  file          = Util.getParam     (params, Params.FNAME);
-		String  scalingDir    = Util.getParam     (params, Params.SCALING_DIR, "width");
-		boolean scaling       = Util.getParam     (params, Params.SCALING, false);
-		int     scalingFactor = Util.getParam     (params, Params.SCALING_FACTOR, 1);
-		
-		boolean createSmall        = Util.getParam(params, Params.CREATE_SMALL,        false);
-		String  smallScalingDir    = Util.getParam(params, Params.SMALL_SCALING_DIR,   "");
-		int     smallScalingFactor = Util.getParam(params, Params.SMALL_SCALING_FACTOR, 0);
-		
-		String dataDir = Lib.resource.getDir(context, Params.Access.PUBLIC, id);
-		
-		if (!new File(dataDir).mkdirs())
-			context.error("Failed to make dir: " + dataDir);
+        String dataDir = createDataDir(id, context);
 		
 		//-----------------------------------------------------------------------
 		//--- create the small thumbnail, removing the old one
+        boolean createSmall        = Util.getParam(params, Params.CREATE_SMALL,        false);
+        String  file          = Util.getParam     (params, Params.FNAME);
+        String  scalingDir    = Util.getParam     (params, Params.SCALING_DIR, "width");
+        boolean scaling       = Util.getParam     (params, Params.SCALING, false);
+        int     scalingFactor = Util.getParam     (params, Params.SCALING_FACTOR, 1);
+        String  type          = Util.getParam     (params, Params.TYPE);
+//        String  version       = Util.getParam     (params, Params.VERSION);
 
-		if (createSmall)
-		{
+        if (createSmall) {
 			String smallFile = getFileName(file, true);
 			String inFile    = context.getUploadDir() + file;
 			String outFile   = dataDir + smallFile;
+            String  smallScalingDir    = Util.getParam(params, Params.SMALL_SCALING_DIR,   "");
+            int     smallScalingFactor = Util.getParam(params, Params.SMALL_SCALING_FACTOR, 0);
 			// FIXME should be done before removeOldThumbnail(context, dbms, id, "small");
 			createThumbnail(inFile, outFile, smallScalingFactor, smallScalingDir);
-			dataMan.setThumbnail(context, dbms, id, true, smallFile);
-		}
+            dataMan.setThumbnail(context, dbms, id, true, smallFile, false);
+       }
 
 		//-----------------------------------------------------------------------
 		//--- create the requested thumbnail, removing the old one
-		// FIXME removeOldThumbnail(context, id, type);
-
-		if (scaling)
-		{
-			String newFile = getFileName(file, type.equals("small"));
-			String inFile  = context.getUploadDir() + file;
-			String outFile = dataDir + newFile;
-
-			createThumbnail(inFile, outFile, scalingFactor, scalingDir);
-
-			if (!new File(inFile).delete())
-				context.error("Error while deleting thumbnail : "+inFile);
-
-			dataMan.setThumbnail(context, dbms, id, type.equals("small"), newFile);
-		}
-		else
-		{
-			//--- move uploaded file to destination directory
-			File inFile  = new File(context.getUploadDir(), file);
-			File outFile = new File(dataDir,                file);
-
-			try {
-				FileUtils.moveFile(inFile, outFile);
-			} catch (Exception e) {
-				inFile.delete();
-				throw new Exception(
-						"Unable to move uploaded thumbnail to destination: " + outFile + ". Error: " + e.getMessage());
-			}
-			
-			dataMan.setThumbnail(context, dbms, id, type.equals("small"), file);
-		}
+        removeOldThumbnail(context,dbms,id,type, false);
+        saveThumbnail(scaling, file, type, dataDir, scalingDir, scalingFactor, dataMan, dbms, id, context);
 
 		//-----------------------------------------------------------------------
-
+		dataMan.indexInThreadPool(context, id, dbms);
 		Element response = new Element("Response");
 		response.addContent(new Element("id").setText(id));
 		// NOT NEEDEDresponse.addContent(new Element("version").setText(dataMan.getNewVersion(id)));
@@ -243,8 +224,9 @@ public class Set implements Service
 	public void addHarvested(Element params, ServiceContext context, Dbms dbms, DataManager dataMan) throws Exception
         {
             String  id            = Util.getParam     (params, Params.ID);
+            String dataDir = createDataDir(id, context);
             String  type          = Util.getParam     (params, Params.TYPE);
-            String  version       = Util.getParam     (params, Params.VERSION);
+//            String  version       = Util.getParam     (params, Params.VERSION);
             String  file          = Util.getParam     (params, Params.FNAME);
             String  scalingDir    = Util.getParam     (params, Params.SCALING_DIR, "width");
             boolean scaling       = Util.getParam     (params, Params.SCALING, false);
@@ -253,10 +235,9 @@ public class Set implements Service
             boolean createSmall        = Util.getParam(params, Params.CREATE_SMALL,        false);
             String  smallScalingDir    = Util.getParam(params, Params.SMALL_SCALING_DIR,   "");
             int     smallScalingFactor = Util.getParam(params, Params.SMALL_SCALING_FACTOR, 0);
-		
-            String dataDir = Lib.resource.getDir(context, Params.Access.PUBLIC, id);
-            if (!new File(dataDir).mkdirs()) context.error("Failed to make dir: " + dataDir);
-		
+
+
+
             //-----------------------------------------------------------------------
             //--- create the small thumbnail, removing the old one
 
@@ -264,44 +245,24 @@ public class Set implements Service
                 String smallFile = getFileName(file, true);
                 String inFile    = context.getUploadDir() + file;
                 String outFile   = dataDir + smallFile;
-                removeOldThumbnail(context,dbms,id,"small");
+                removeOldThumbnail(context,dbms,id,"small", false);
                 createThumbnail(inFile, outFile, smallScalingFactor, smallScalingDir);
-                dataMan.setThumbnail(context, dbms, id, true, smallFile);
+                dataMan.setThumbnail(context, dbms, id, true, smallFile, false);
             }
 
             //-----------------------------------------------------------------------
             //--- create the requested thumbnail
             
-            removeOldThumbnail(context,dbms,id,type);
+            removeOldThumbnail(context,dbms,id,type, false);
+            saveThumbnail(scaling, file, type, dataDir, scalingDir, scalingFactor, dataMan, dbms, id, context);
 
-            if (scaling) {
-                String newFile = getFileName(file, type.equals("small"));
-                String inFile  = context.getUploadDir() + file;
-                String outFile = dataDir + newFile;
-                
-                createThumbnail(inFile, outFile, scalingFactor, scalingDir);
-                if (!new File(inFile).delete()) context.error("Error while deleting thumbnail : "+inFile);
-                dataMan.setThumbnail(context, dbms, id, type.equals("small"), newFile);
-            } else {
-                //--- move uploaded file to destination directory
-                File inFile  = new File(context.getUploadDir(), file);
-                File outFile = new File(dataDir,                file);
-
-                try {
-                    FileUtils.moveFile(inFile, outFile);
-                } catch (Exception e) {
-                    inFile.delete();
-                    throw new Exception("Unable to move uploaded thumbnail to destination: " + outFile + ". Error: " + e.getMessage());
-                }
-			
-                dataMan.setThumbnail(context, dbms, id, type.equals("small"), file);
-            }
+            dataMan.indexInThreadPool(context, id, dbms);
         }
         
         public void removeHarvested(Element params, ServiceContext context, Dbms dbms) throws Exception {
             String  id   = Util.getParam(params, Params.ID);
             String  type = Util.getParam(params, Params.TYPE);
-            removeOldThumbnail(context,dbms,id,type);
+            removeOldThumbnail(context,dbms,id,type, true);
         }
 
 	//--------------------------------------------------------------------------
@@ -310,7 +271,48 @@ public class Set implements Service
 	//---
 	//--------------------------------------------------------------------------
 
-	private void removeOldThumbnail(ServiceContext context, Dbms dbms, String id, String type) throws Exception
+    /**
+     * TODO Javadoc.
+     *
+     * @param scaling
+     * @param file
+     * @param type
+     * @param dataDir
+     * @param scalingDir
+     * @param scalingFactor
+     * @param dataMan
+     * @param dbms
+     * @param id
+     * @param context
+     * @throws Exception
+     */
+    private void saveThumbnail(boolean scaling, String file, String type, String dataDir, String scalingDir,
+                               int scalingFactor, DataManager dataMan, Dbms dbms, String id, ServiceContext context) throws Exception {
+            if (scaling) {
+                String newFile = getFileName(file, type.equals("small"));
+                String inFile  = context.getUploadDir() + file;
+                String outFile = dataDir + newFile;
+                
+                createThumbnail(inFile, outFile, scalingFactor, scalingDir);
+                if (!new File(inFile).delete()) context.error("Error while deleting thumbnail : "+inFile);
+                dataMan.setThumbnail(context, dbms, id, type.equals("small"), newFile, false);
+            } else {
+                //--- move uploaded file to destination directory
+                File inFile  = new File(context.getUploadDir(), file);
+                File outFile = new File(dataDir,                file);
+
+                try {
+                    FileUtils.moveFile(inFile, outFile);
+                } catch (Exception e) {
+                    IO.delete(inFile, false, context);
+                    throw new Exception("Unable to move uploaded thumbnail to destination: " + outFile + ". Error: " + e.getMessage());
+                }
+			
+                dataMan.setThumbnail(context, dbms, id, type.equals("small"), file, false);
+            }
+        }
+
+	private void removeOldThumbnail(ServiceContext context, Dbms dbms, String id, String type, boolean indexAfterChange) throws Exception
 	{
 		GeonetContext gc = (GeonetContext) context.getHandlerContext(Geonet.CONTEXT_NAME);
 
@@ -331,7 +333,7 @@ public class Set implements Service
 		//-----------------------------------------------------------------------
 		//--- remove thumbnail
 
-		dataMan.unsetThumbnail(context, dbms, id, type.equals("small"));
+		dataMan.unsetThumbnail(context, dbms, id, type.equals("small"), indexAfterChange);
 
 		//--- remove file
 
@@ -418,13 +420,19 @@ public class Set implements Service
 
 	private Image getTiffImage(String inFile) throws IOException
 	{
-		Tiff t = new Tiff();
-		t.readInputStream(new FileInputStream(inFile));
-
-		if (t.getPageCount() == 0)
-			throw new IOException("No images inside TIFF file");
-
-		return t.getImage(0);
+	    FileInputStream fileInputStream = null;
+	    try {
+    		Tiff t = new Tiff();
+            fileInputStream = new FileInputStream(inFile);
+            t.readInputStream(fileInputStream);
+    
+    		if (t.getPageCount() == 0)
+    			throw new IOException("No images inside TIFF file");
+    
+    		return t.getImage(0);
+	    } finally {
+	        IOUtils.closeQuietly(fileInputStream);
+	    }
 	}
 
 	/**
@@ -454,7 +462,3 @@ public class Set implements Service
 	private static final String FNAME_PARAM   = "fname=";
 
 }
-
-//=============================================================================
-
-
